ML-model/requirements.txt
nats-py==2.3.1
pandas==2.1.1
joblib==1.3.2
scikit-learn==1.2.2
tabpfn==0.1.9